backend/requirements.txt
blinker==1.9.0
boto3==1.40.1
botocore==1.40.1
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
PyJWT>=2.0
python-dotenv==1.0.1
